[refactor] 將 routes/users.js 的 handlers 抽出為 userController.js
